squashing warnings

+++ src/jsutils.c
@@ -468,7 +468,7 @@ unsigned int rand_m_z = 0xCAFEBABE;    /* must not be zero */
 int rand() {
   rand_m_z = 36969 * (rand_m_z & 65535) + (rand_m_z >> 16);
   rand_m_w = 18000 * (rand_m_w & 65535) + (rand_m_w >> 16);
-  return RAND_MAX & (int)((rand_m_z << 16) + rand_m_w);  /* 32-bit result */
+  return (int)RAND_MAX & (int)((rand_m_z << 16) + rand_m_w);  /* 32-bit result */
 }
 
 void srand(unsigned int seed) {
